refactor: spectra analysis components paths
