+++ src/component/panels/multipleAnalysisPanel/AlignSpectra.tsx
