+++ src/component/panels/multipleAnalysisPanel/AnalysisChart.tsx
