+++ src/component/panels/multipleAnalysisPanel/MultipleSpectraAnalysisPanel.tsx
