+++ src/component/panels/multipleAnalysisPanel/MultipleSpectraAnalysisTable.tsx
