+++ src/component/panels/multipleAnalysisPanel/base/AnalysisCell.tsx
