+++ src/component/panels/multipleAnalysisPanel/base/AnalysisColumnHeader.tsx
