+++ src/component/panels/multipleAnalysisPanel/preferences/AnalysisTablePreferences.tsx
