+++ src/component/panels/multipleAnalysisPanel/preferences/LegendsPreferences.tsx
