+++ src/component/panels/multipleAnalysisPanel/preferences/index.tsx
